+++ test/fixtures/cookbooks/test/recipes/plugins.rb
@@ -5,8 +5,6 @@ grafana_config_auth 'Grafana' do
   anonymous_enabled true
 end
 
-grafana_config_writer 'Grafana'
-
 grafana_plugin 'grafana-clock-panel' do
   action :install
 end
@@ -19,4 +17,7 @@ end
 # Tests are failing as the server has not fully become available when tests run
 chef_sleep 'Sleep so inspec tests pass' do
   seconds 25
+end
+grafana_service 'grafana' do
+  action %i(enable start)
 end