+++ test/fixtures/cookbooks/test/recipes/proxy.rb
@@ -2,7 +2,10 @@ auth_header = 'X-WEBAUTH-USER'
 
 grafana_install 'grafana'
 
-grafana_config 'Grafana'
+grafana_config 'Grafana' do
+  # In test we turn of sensitive so we can get better logs
+  sensitive false
+end
 
 grafana_config_auth 'Grafana' do
   disable_login_form true
@@ -13,7 +16,10 @@ grafana_config_auth 'Grafana' do
   login_cookie_name 'grafana_session'
 end
 
-grafana_config_writer 'Grafana'
+grafana_service 'grafana' do
+  action %i(enable start)
+  delay_start false
+end
 
 grafana_user 'j.smith' do
   auth_proxy_header auth_header