+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file is used to list changes made in each version of grafana.
 
 ## Unreleased
 
+- Set resource configuration Hash programmatically
+- Automatic config Hash path creation
+- Automatic resource property fetching
+- Automatic config template resource generation
+- Add service resource
+- Add LDAP log filters to the main config file
+
 ## 9.6.2 - *2021-06-01*
 
 ## 9.6.1 - *2021-05-11*